init refactor of creation sub-components, external styles, and creation actions
