+++ src/components/Creation/CreationCard/CreationSocials/CreationSocialsStyles.tsx
@@ -1,0 +1,39 @@
+// STYLES
+import styled from 'styled-components'
+
+const CreationSocialsStyles = styled.div`
+  display: flex;
+  align-items: center;
+  // background: lime;
+  justify-content: flex-end;
+
+  .cr-socials-main {
+    display: flex;
+    // background: red;
+    align-items: center;
+    // height: 50px;
+    // border: 2px solid black;
+  }
+  .cr-social {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: ;
+  }
+  .cr-social .btn {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+  }
+  .cr-social .text {
+    color: white;
+  }
+  .cr-social .icon {
+    // color: white;
+    height: 15px;
+    // background: green;
+  }
+`
+
+export default CreationSocialsStyles
